app/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
pydantic
requests 